requirements.txt
numpy==1.19.2
Pillow==8.0.0
streamlit==0.69.2
